Refactor skylink functionality
