+++ src/index.ts
@@ -1,16 +1,19 @@
 export { SkynetClient } from "./client";
 export { deriveChildSeed, genKeyPairAndSeed, genKeyPairFromSeed } from "./crypto";
-export { signEntry } from "./registry";
+export { getSkylinkUrlForPortal } from "./download";
+export { getEntryUrlForPortal, signEntry } from "./registry";
 export { getRelativeFilePath, getRootDirectory } from "./utils/file";
 export { MAX_REVISION } from "./utils/number";
-export { parseSkylink, uriHandshakePrefix, uriHandshakeResolverPrefix, uriSkynetPrefix } from "./utils/skylink";
+export { convertSkylinkToBase32 } from "./skylink/format";
+export { parseSkylink } from "./skylink/parse";
 export {
   defaultPortalUrl,
   defaultSkynetPortalUrl,
   extractDomainForPortal,
   getFullDomainUrlForPortal,
-  getEntryUrlForPortal,
-  getSkylinkUrlForPortal,
+  uriHandshakePrefix,
+  uriHandshakeResolverPrefix,
+  uriSkynetPrefix,
 } from "./utils/url";
 export { DacLibrary, mySkyDomain, mySkyDevDomain } from "./mysky";
 // Re-export Permissions.
@@ -34,4 +37,4 @@ export type { CustomConnectorOptions, MySky } from "./mysky";
 export type { CustomGetEntryOptions, CustomSetEntryOptions, SignedRegistryEntry, RegistryEntry } from "./registry";
 export type { CustomGetJSONOptions, CustomSetJSONOptions, JsonData, JSONResponse } from "./skydb";
 export type { CustomUploadOptions, UploadRequestResponse } from "./upload";
-export type { ParseSkylinkOptions } from "./utils/skylink";
+export type { ParseSkylinkOptions } from "./skylink/parse";
