+++ src/skylink/index.ts
@@ -1,0 +1,11 @@
+export class SiaSkylink {
+  toBytes(): Uint8Array {
+    // TODO
+  }
+
+  toString(): string {
+    // TODO
+  }
+}
+
+export function newSkylinkV2(siaPublicKey: SiaPublicKey, tweak: Uint8Array): SiaSkylink {}
