+++ src/skylink/parse.ts
@@ -1,8 +1,7 @@
-import base64 from "base64-js";
-import base32Encode from "base32-encode";
 import parse from "url-parse";
-import { trimForwardSlash, trimSuffix, trimUriPrefix } from "./string";
-import { validateOptionalObject, validateString } from "./validation";
+import { trimForwardSlash, trimSuffix, trimUriPrefix } from "../utils/string";
+import { uriSkynetPrefix } from "../utils/url";
+import { validateOptionalObject, validateString } from "../utils/validation";
 
 /**
  * Parse skylink options.
@@ -22,37 +21,6 @@ const defaultParseSkylinkOptions = {
   includePath: false,
   onlyPath: false,
 };
-
-export const uriHandshakePrefix = "hns:";
-export const uriHandshakeResolverPrefix = "hnsres:";
-export const uriSkynetPrefix = "sia:";
-
-/**
- * Converts the given base64 skylink to base32.
- *
- * @param skylink - The base64 skylink.
- * @returns - The converted base32 skylink.
- */
-export function convertSkylinkToBase32(skylink: string): string {
-  const decoded = base64.toByteArray(skylink.padEnd(skylink.length + 4 - (skylink.length % 4), "="));
-  return base32Encode(decoded, "RFC4648-HEX", { padding: false }).toLowerCase();
-}
-
-/**
- * Formats the skylink by adding the sia: prefix.
- *
- * @param skylink - The skylink.
- * @returns - The formatted skylink.
- */
-export function formatSkylink(skylink: string): string {
-  if (skylink == "") {
-    return skylink;
-  }
-  if (!skylink.startsWith(uriSkynetPrefix)) {
-    skylink = `${uriSkynetPrefix}${skylink}`;
-  }
-  return skylink;
-}
 
 const SKYLINK_MATCHER = "([a-zA-Z0-9_-]{46})";
 const SKYLINK_MATCHER_SUBDOMAIN = "([a-z0-9_-]{55})";
